Added stuff from Numba to make it faster
I imported vectorize from numba to make things more optimized.
All the methods currently are pretty good in terms of speed, but
the backpropagation method is the bottle neck of this model. I have tried
different ways of making it a little faster but this is the best so far...